+++ src/core/ExocoreGateway.sol
@@ -15,13 +15,13 @@ import {
 } from "../lzApp/OAppUpgradeable.sol";
 import {ExocoreGatewayStorage} from "../storage/ExocoreGatewayStorage.sol";
 
+import {OAppCoreUpgradeable} from "../lzApp/OAppCoreUpgradeable.sol";
+import {IOAppCore} from "@layerzero-v2/oapp/contracts/oapp/interfaces/IOAppCore.sol";
 import {OptionsBuilder} from "@layerzero-v2/oapp/contracts/oapp/libs/OptionsBuilder.sol";
 import {ILayerZeroReceiver} from "@layerzero-v2/protocol/contracts/interfaces/ILayerZeroReceiver.sol";
 import {OwnableUpgradeable} from "@openzeppelin-upgradeable/contracts/access/OwnableUpgradeable.sol";
 import {Initializable} from "@openzeppelin-upgradeable/contracts/proxy/utils/Initializable.sol";
 import {PausableUpgradeable} from "@openzeppelin-upgradeable/contracts/utils/PausableUpgradeable.sol";
-import {IOAppCore} from "@layerzero-v2/oapp/contracts/oapp/interfaces/IOAppCore.sol";
-import {OAppCoreUpgradeable} from "../lzApp/OAppCoreUpgradeable.sol";
 
 contract ExocoreGateway is
     Initializable,
@@ -117,11 +117,11 @@ contract ExocoreGateway is
      * @dev Indicates that the peer is trusted to send LayerZero messages to this OApp.
      * @dev Peer is a bytes32 to accommodate non-evm chains.
      */
-    function setPeer(uint32 clientChainId, bytes32 clientChainGateway) 
-        public 
-        override(IOAppCore, OAppCoreUpgradeable) 
-        onlyOwner 
-        whenNotPaused 
+    function setPeer(uint32 clientChainId, bytes32 clientChainGateway)
+        public
+        override(IOAppCore, OAppCoreUpgradeable)
+        onlyOwner
+        whenNotPaused
     {
         require(clientChainId != uint32(0), "ExocoreGateway: zero value is not invalid endpoint id");
         require(clientChainGateway != bytes32(0), "ExocoreGateway: client chain gateway can not be empty");
@@ -152,13 +152,16 @@ contract ExocoreGateway is
         }
     }
 
-    function requestRegisterTokens(uint32 srcChainId, uint64 lzNonce, bytes calldata payload) public onlyCalledFromThis {
+    function requestRegisterTokens(uint32 srcChainId, uint64 lzNonce, bytes calldata payload)
+        public
+        onlyCalledFromThis
+    {
         uint8 count = uint8(payload[0]);
         uint256 expectedLength = count * ADDRESS_LENGTH + 1;
         _validatePayloadLength(payload, expectedLength, Action.REQUEST_DEPOSIT);
 
         bytes[] memory tokens = new bytes[](count);
-        for (uint i; i < count; i++) {
+        for (uint256 i; i < count; i++) {
             uint256 start = i * ADDRESS_LENGTH + 1;
             uint256 end = start + ADDRESS_LENGTH;
             tokens[i] = payload[start:end];
@@ -344,4 +347,5 @@ contract ExocoreGateway is
             revert UnexpectedInboundNonce(inboundNonce[srcEid][sender], nonce);
         }
     }
+
 }
