+++ test/mocks/AssetsMock.sol
@@ -3,6 +3,7 @@ pragma solidity ^0.8.19;
 import {IAssets} from "src/interfaces/precompiles/IAssets.sol";
 
 contract AssetsMock is IAssets {
+
     address constant VIRTUAL_STAKED_ETH_ADDRESS = 0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE;
 
     mapping(uint32 => mapping(bytes => mapping(bytes => uint256))) public principleBalances;
@@ -37,7 +38,7 @@ contract AssetsMock is IAssets {
         if (bytes32(assetsAddress) != bytes32(bytes20(VIRTUAL_STAKED_ETH_ADDRESS))) {
             require(isRegisteredToken[clientChainLzId][assetsAddress], "the token is not registered before");
         }
-        
+
         require(opAmount <= principleBalances[clientChainLzId][assetsAddress][withdrawer], "withdraw amount overflow");
 
         principleBalances[clientChainLzId][assetsAddress][withdrawer] -= opAmount;
@@ -60,7 +61,7 @@ contract AssetsMock is IAssets {
     function registerTokens(uint32 chainId, bytes[] memory tokens) external returns (bool) {
         require(isRegisteredChain[chainId], "the chain is not registered before");
 
-        for (uint i; i < tokens.length; i++) {
+        for (uint256 i; i < tokens.length; i++) {
             bytes memory token = tokens[i];
             require(token.length == 32, "token address with invalid length");
             require(!isRegisteredToken[chainId][token], "already registered token");
